POPUPS YAAAAAAAY
I added a couple of popups. The first will display a warning message if
the user attempts to enter a page out of the range of the book into the
textbox. The second will display whenever the user decides to either
close the file or to open a new comic when one is already displayed. It
asks whether the program should bookmark their page for them.